reports/accounting: remove LRU caching of output
Because of the changes in I4bb6a1487a8f0f4412953a26882bac268503886c, the
output of the Google Sheets API is cached in an LRU cache, not just
between different tests, but also between subsequent runs of the report.

This causes lag between updates to the spreadsheet and their output in
the report, which may be confusing to users.  Moreover, because of the
report currently having some formulas that take a while to resolve, this
results into confusing output with "#NAME?" or "Loading...". Previously,
it would clear up with a second run (or with running these in an often
enough cadence, e.g. from Icinga).

Partially rollback the changes of the changeset above and remove the LRU
cache; instead, fix the underlying issue that the changeset was trying
to fix, by overloading the run() method, rather than the __init__()
method. Judging by the Netbox report code, run() is called once at each
run of the report, and is responsible for calling all the test methods
one by one, so it feels appropriate to run asset gathering there.

Change-Id: I91aae093347d3854601de76a4077698441fa0c70