reports/accounting: avoid evaluating formulas
The spreadsheet that the report is polling, is often using a custom
spreadsheet JavaScript formula (NETBOX_BYSERIAL), which polls the Netbox
API to resolve the serial number -> asset tag mapping. This is often
used for the most recent month to avoid toil around manual input, while
the historical data is typically kept as static strings (usually with a
copy/paste values only routine).

In turn, this means that this Netbox report polling this spreadsheet is
generating a lot of API requests to Netbox itself, amplifying the cost
of the report run, while also pointlessly checking whether its own API
output matches its own database. This is also brittle due to the way
GSheets works, and often results into the output being #NAME? or
"Loading..." if the spreadsheet cache is cold.

To resolve this, pass valueRenderOption="FORMULA" to the API, to avoid
rendering any formulas, and then explicitly ignore formulas starting
with "NETBOX". This has the side-effect of rendering dates differently
(using Lotus 1-2-3 format), so pass dateTimeRenderOption to revert to
the previous behavior.

Note that this WILL BREAK any other formulas that this spreadsheet may
have in the future, including simple ones (e.g. =1+2). If that becomes
desirable in the future, this commit can be reverted at that time.

Change-Id: Ibb658771a84e732436652e55d981971f17b149ff