interface_automation: Make device selection a text field
This fixes the UI problem of having to select the device from a gigantic list of devices, instead
allowing the device to be specified by name in a free text field. Additionally it allows multiple
devices to be selected at once, each processed in turn.

This also fixes a minor bug where an additional IP address would be allocated to mgmt interfaces
even if one already existed, making it safe to rerun on machines without first checking if they
are set up.

Change-Id: I1dbb8e2c8b7a057adbd5046bcc8ec2885650a6ac