interface_automation.py: Minor refactors and fixes for 2.9
This fixes the non-shared portions of interface creation by moving it
to a separate method based on if it is operating on a virtual machine
or not. It also fixes a minor differentiating issue in a closely
shared bit of code by fencing out the bit that is not possible on a
VMInterface and changing the log message.

Bug: T266487
Change-Id: Ice047f9b460faaddfb3b2961d8f0ef66dd7c7de0