Cleanup new interface creation and add logic to remove orphan ints
Few cleanup bits here and some logic to prrevent an issue which can
occur due to how Juniper devices rename interfaces based on their
speed, which can change if the same port supports multiple PHYs and
the module is replaced.

I've been able to test this on netbox-next and I'm satisfied it is
working ok in all the cases where the new code is executed.
Specifically:

1. For PuppetDB Import
2. Provisioning a new server initially
3. Moving a server within a row

I've documented the test results for all 3 in the phab task, please
feel free to review there and let me know if there are any issues
with the way it's working.  Looks ok to my eyes.

EDIT: Latest patchset also logs error if the "orphan" interface still
has an IP address attached, as well as in case where it has cable.

Also some input validation added to ensure interface names start with
valid Juniper interface prefix, and if adding to a switch in a
virtual-chassis that the first digit matches the VC ID.

Bug: T301392
Change-Id: I004fe661d5b074357aa8041a7ff0fdb91538ce90