Update Netbox Move Server Script to Copy original Tagged Vlans
The current move server script only configures the new switch port
with the primary, untagged Vlan in use on the link, and sets the
switch side to access.

These changes adjust it so if there are tagged vlans on the old
interface it will copy them to the new one, and set the mode on the
new one to tagged also.

Change-Id: I780eeff40b8be1407fded2a1621632d9a4df5b45