Change order that Netbox server provision script gets old/new vlan name
The provision script is built to search for two variants of naming
convention for vlans, an "old" and a "new" variant.

The 'new' variant assumes a 'vlan per rack' model in which the rack
ID is part of the vlan name.  The 'old' variant either assumes the
vlan name is cloud-hosts1-<site> for cloud hosts, or uses
<vlan_type>1-<row>-<site>, for the row-wide vlans.

Currently it requests the 'old' name first, and if this is not found
in Netbox tries to find the 'new' one.  Because of the way the name
defaults for cloud hosts, in Eqiad it always manages to find
'cloud-hosts1-eqiad', and never goes on to find
'cloud-hosts1-e4-eqiad' or similar.  When it defaults to the first
the script then fails as the vlan returned is not present on the
switch in question.

This patch changes the order of this search so it requests the 'new'
naming format first, and then the 'old'.  This should be safe as the
'new' naming is more specific and therefore should be used if
present.

Bug: T304936
Change-Id: Ibc4a46fc267b24c69286aa1f6861f21c0b18caa9