Add check to network report to ensure IPs match connected Vlans
The test_matching_vlan() function goes through all server interfaces
with a cable attached and IPs configured, and checks that the IPs
are part of the correct subnet for the Vlan the far-side switch port
is connected to.

It uses get() to pull the prefix for each address family for the Vlan,
in a try/except block so that on failure it will log the fact that
the Vlan has more than one prefix attached in a given address family.

Tested on netbox-next (which has some failures as it's data is out of
date). Example output:

https://phabricator.wikimedia.org/F35248856

Bug: T310299
Change-Id: I481eead74daa547f53aae79b724399e0907b72c5