Fix Network report to deal with IP address on host with no Vlan pfx
When I pushed the Network report with the test_matching_vlan()
function it threw a DoesNotExist exception getting the IPv6 prefix
for an attached Vlan.

In our infra this shouldn't happen, as all host-facing Vlans should
have both IPv4 and IPv6 prefixes attached.  The only exceptions to
this are in the WMCS cloud realm, which is IPv4 only.  However, as
we do not track the server IP allocations for those prefixes in
Netbox, the check is not run for those Vlans.

In the case of the error one of those interfaces had an IP bound to
it in error, which caused the test to query for the attached Vlan's
prefixes, and triggered when it found none for IPv6.

This change catches the exception, should it happen, and issues a
warning.

Change-Id: Ib997ed7471925188edc827f3cdb311f4e2fa73c8