Set switch-side MTU to 9192 for discovered links from servers
It seems that for servers with 2 ports, for which the second link is
added to Netbox by the Pupept import script, the MTU was not getting
set on the switch-side port.

This patch should ensure it's added at this stage.  I've tested for an
example host in netbox-next and it works as expected.  I believe it's
safe, only time it should run is when connecting to a L2 switch.

Bug: T329799
Change-Id: I296f5049ea88c86d4f1da08f762826589126b188