Do not compare speed of disabled interfaces when validating blocks
Quick fix so our netbox report doesn't fire if an interface that
exists, but is not enabled, does not match the speed of the other
interfaces in the same block on QFX5120 devices.

Should be moved to a validator but this will stop the report firing
unnecessarily as it did today.

Bug: T303529
Change-Id: I153b1b54481dd933735620c7e92e8cbdb2ee84e1