Fail when setting int relations if PuppetDB parent not found in Netbox
Adds a check to fail if Puppet facts list an interface as a parent
of the one we are processing, but the referenced parent is not found
in Netbox.

_get_ordered_ints() is designed to prevent this from happening,
ensuring we process bridge and physical interfaces before any
children are handled.

That's worked fine up until now, and logic seems sound, but we
recently we got an unhandled excpetion in _update_int_relations()
for this reason.  How exactly it happened is not clear, but at
least now we'll fail cleanly, and the error should provide a
clure.

Also some changes to comments/docstrings for clarity.

Tested on netbox-next with against a selection of hosts and no
problems encountered.

Bug: T350479
Change-Id: I5f8ecff8ed3319e4e0e98324ee36d16e7ae793d6