Change move server script to skip devices already connected to new sw
On previous runs the move server script would error (see
https://phabricator.wikimedia.org/P56775), after trying to operate
on a server that was already connected to the new switch.

The simplest fix I think is to just not process any servers already
connected to the new switch (be they pre-migrated or newly deployed
since the new switches became operational).

Tested on netbox-next.

Bug: T348129
Change-Id: I6805dc2f5cb3e233b299a7830ad8cc1a49d836a3