Fix error when removing an interface's bridge membership
This bug was hit after some non-standard order of operations whereby
a host interface previously had a relationship (parent link or bridge
membership), that is no longer there.

The code was written to allow for this, settting bridge to 'none' if
it's got an existing value and puppetdb no longer shows any bridge
member.  However an error in the logging tried to get 'parent.name'
which fails if that parent is None.  Changing to use just 'parent'
in the log message (which will return the .name as it's being
cast to string).

Thanks @ayounsi for the task and correctly identifyinig the issue :)

Bug: T359629
Change-Id: I6d1a35da05e5606b093f51c29e2f7a40c24723df