Netbox custom script to add additional IPv4 addresses to host
Adds additional IPv4 addresses to a host's primary interface, and
gives each a DNS name with '-a', '-b', '-c' appended to the first
part to differentiate.  The use-case if for Cassandra hosts,
which currently need such IPs / dns records.

As discussed previously the goal here is to separate the addition
of second and subsequent IPs for Cassandra hosts from the Provision
script, so we don't need to specify the requirement to DC-Ops ahead
of the provision stage.

Bug: T358096
Change-Id: I755d072a637c605273e46257b6d4abff7d13202b