Validate IRB interface names correspond to vlan and refactor
Adds a check to validate that any 'irb' interface name has a valid
vlan-id after the dot.  Protects against a typo that would create
the irb but not match.

Also refactor the file wrapping the existing tests in an 'if'
checking if the device is in NETWORK_ROLES.  Previously this was
being checked separately under each test, but seems better to only
do once?

I've tested all the combinations on netbox-next and seems to work ok,
however we seem to be hitting the same (bug?) as in change
I07840894b0c0c41d52b7b738299e04c0a23612e3.  Namely the validation for
the irb stuff works, but it does not return the correct error from
the associated self.fail() call when it does.  Given it's mostly
netops who would create / modify such elements I think we can maybe
merge anyway, as we are aware of the problem.

Bug: T366348
Change-Id: I8b372f6a08cc1f1c7c3bf570c77ef20f3fe6407c