Only validate vlan location in provision script when manually selected
A quirk in the location validation meant it was failing in codfw
racks with new per-rack vlans, as these vlans are trunked in
remote rows to LVS but nowhere else. TL;DR confusing the logic.

This changes the approach to only validate the vlan is valid for a
location if it has been manually selected by the user.  If the
vlan has been selected based on type + naming pattern (the
typical workflow) then we assume the correct one has been chosen
for the given rack and don't re-check.

Bug: T372654
Change-Id: I67ccbe38cd5c7de2bb0a187bbbca9085dded3ab7