Allow the selection of any vlan in provision server script
Currently the ProvisionServer script allows the user to manually
specify a vlan to add a server to.  However it filters the
presented options to only those which have names not matching
our standard 'types' that can be selected in the type dialog.

That was ok traditionally, but right now for instance it prevents
selecting private1-d-codfw as a vlan, and using the 'type' drop-down
for a vlan in codfw row D will select a per-rack vlan like
'private1-d2-codfw'.  Mostly that's fine but for some cases like
Ganeti we may need to still add to the row-wide one.

This patch removes the filter so all vlans can be selected.  It would
be nice to filter to only those on the site/valid but that is not
possible when the form is displayed.  The _is_vlan_valid() function
is still called afterwards which will prevent a user from selecting
a vlan from the wrong site or row etc.

Bug: T365651
Change-Id: I465d4fd53efc69f9e79c29512035ce64ef6ee2fe