Provision script: don't ask the user for v6 AAAA
Instead do it automatically based on a (now) short list of hosts.
A bit similar to what we do in the Network report, but without the
prefixes of hosts that now have v6 (eg. v6 added during refresh).

This will eliminate operator error and simplify the workflows further.

Also add a few types.

Change-Id: If9d3d18176048d41b63e9982bc5656ff0d82d4e1