Fix some bugs with the move_server Netbox script
Fixes some issues discovered using this since the upgrade to Netbox 4

1) Now rack postitions are floats, as you can have something at
height 1.5 if you want.  This meant the new interfaces were
getting created with ".0" appended at the end.  Cast to int
first to avoid this, we have no or are unlikely to ever have
something racked at a fraction of a U so this should be fine.

2) Adjust how the vlans are passed to the _update_z_nbiface()
function in common.py, passing a list of Netbox vlan object ids. In
trial-and-error this was the only way the tagged vlans could be
made to save.  Also changed the update_z_nbiface() function in
common.py to change the order of operations, including an extra
save() that shouldn't be needed, but is if we want it to properly
set the tagged_vlans.

Tested and working as expected on netbox-next.

Bug: T374024
Change-Id: Ifdd8cbcfa4fca17266da896835bd695da6079d1c