Prefix validator: ensure k8s role and site
If creating a new prefix that is a child of a prefix with the kubernetes
roles, ensure that the new prefix also have the kubernetes role and the
same site as parent.

Also disable pylint too-many-positional-arguments as it got introduced
in a recent CI update and fails for unrelated parts of the code.

Tested on netbox-next.

Bug: T354169
Change-Id: I7bd569587d28c9ef11db853c68bfa50be4888c06