Only try to find 'real' netmask for IPs if they are /32 or /128
The logic for certain classes of machines which are 'vip_exempt' to
get the _real_ netmask for IPs was causing an issue when the actual
server primary IP was passed.  In that case it was finding the
parent prefix of the real vlan subnet, like a /12 or something.

We don't need to find the real netmask for IPs which already have
it, so we can change this line to only try if the IP in question
has a /32 or /128 prefixlen.

Bug: T378751
Change-Id: Iea2c2f2202c927f1acce86116716f2cbffbcfba1