Add validators for Netbox IPsec elements
Enforces some agreed on norms which matches what our automation is
configured to allow.  i.e. prevents some IPsec configurations to
be defined in Netbox that are not compatible with our current
automation.

In general forces AES in GCM mode which is more efficient, and checks
that no authentication algo is specififed as well as a separate one
is not needed in that case.

Also makes sure PFS is enabled and that the DH groups used for it and
IKE are at least 14 (2408 bit modp or EC groups).

Bug: T378020
Change-Id: Ic4b8d7d1784715b8574439d6ef8c6e376b714bb2