Fr-tech provision script to assign IPs and switch ports
This script can be used by dc-ops when adding new fundraising
hosts to the datacentre.  The advantage over the current manual
process is:

- It automatically assigns IP addresses (mgmt and prod) and assigns
  valid dns_names for them, so Netbox can drive the DNS.
- It configures the switch ports on both fr-tech L2 switches at the
  specific site for the correct vlan in Netbox, so it can be managed
  using automation.
- It runs the validation code to ensure port selection conforms to
  the restriction we have with the new Trident-3 based switch models.

The server physical ports just use dummy names PRIMARY_A and
PRIMARY_B, which provides the objects to connect to the switches. Both
are made part of the 'bond0' LAG which is where the assigned IP is
attached.

NOTE: Also minor changes to network report and revert.py to make CI
happy.

Bug: T379553
Change-Id: Ie0a0be542eb658dc21dbf35983d382f7281d45fa