Validators: Allow an interface to be called just "irb" on a device
There is a check in the interface name validator to ensure that irb
sub-interface names match actual Vlans we are using.

For instance an interface "irb.101", the gateway for Vlan 101, would
be invalid if we had no Vlan 101 in Netbox.

For SR-Linux, given how the config works, we will need to create a
parent interface called just "irb" for the Nokia devices in Netbox.
Right now this validator fails in that case as it is expecting
anything starting "irb" to be in the format "irb.<integer>".

Patch changes the validation so it is only done for the subints.
Tested on netbox-next and works as expected.

Bug: T371088
Change-Id: I433f13f4afcd6f473d3a1c3deaf363a02ac602f7