Run validator when creating switch-interface in provision script
During the work on the Funraising provision script we discovered that
the Netbox validators are not executed when running custom scripts or
from nbshell itself.

For the most part this is not an issue - we do not have scripts that
add objects with parameters that conflict with our validators.  For
switch ports, however, we very much do want the validator to run and
ensure that each block of 4 ports on Trident-3 based switches is
compatible.  Failing to do it at this early stage is goinig to lead
to messy manual cleanup.

Bug: T383915
Change-Id: I536052caf205fe863b69e367b125b918d0662c7a