Ensure child interfaces of physicals are removed before physical
We had an edge-case where an lvs host was reimaged with the
following changes to the interface layout after it came back up:

- The physical interfaces were renamed as bookworm defaults to using
ID_NET_NAME_SLOT rather than ID_NET_NAME_PATH for names.

- The vlan sub-interfaces were all removed from the device, as
Liberica is only using tunnelled IPIP transport, and doesn't need
them.

The result of this is when the script processed vlan interfaces it
did not modify the existing ones for this host.  Then at the end
when it was deleting any interfaces not found in PuppetDB it hit
an error, trying to delete() an interface that still had the old
vlan child ints attached.

This patch modifies the behaviour to first iterate over all child
interfaces and remove them if they are no longer in puppetdb,
before processing the rest of the interfaces.

Tested on netbox-next.

Bug: T388770
Change-Id: Ia3f33a442ac0ecbb27fb82ee3ddf1c6133b922a6