Update provision and interface validator to support Nokia TOR
This patch makes a few changes to support using Nokia devices as
top-of-rack switches in future.

1) We pass the port-type (i.e. '10gbase-x-sfpp') to the
_update_z_nbiface() function (in common.py), and adjust the
function so that it will change the type of an existing interface
if it doesn't match.  This was not being done for Juniper, as the
interface name provided the speed ('ge/xe/et') and we were creating
new interfaces any time it changed.  The Nokia interface names are
constant so we instead need to pass this and update where needed.

2) Updates the interface validator to ensure that port speeds are
compatible, following the rather odd Nokia grouping (see the image
on https://netbox.wikimedia.org/dcim/device-types/344/)

3) Ensures the interface validator is run after if we try to change
an interface type in _update_z_nbiface(), so if a type is requested
that's incompotaible with the others in the same block the action
will fail.

Tested on netbox-next and working as expected:
https://phabricator.wikimedia.org/P84202

Bug: T405637
Change-Id: Id0d8db0252ab74d04169981e1f10c18495cc5708