Refactor of move_server and script to move selective hosts
The main goal of this patch is to add a new script (under the
'move_server' module) to allow us to move selective server
uplinks from one top-of-rack switch to another.

This builds on the existing code, but moves most of it to the
common.py file, as we do the same thing for the 'move all servers
in a rack' and the new 'move specific servers' scripts.  It also
adds some improvements for instance copying cable colours and types.

In testing it works well and as expected (tested with numerous
hosts on netbox-next):

https://phabricator.wikimedia.org/P84961
https://phabricator.wikimedia.org/P84959

You can try here:

https://netbox-next.wikimedia.org/extras/scripts/88/

It also fixes an issue encountered in the port-block/speed validation,
where we were matching exact types, as that failed in one of the
tests.  Given the exact type doesn't need to match, only the speed, we
extract the speed from the potential types and compare that.  Refactored a few var names too for readability.

Bug: T405640
Change-Id: Ib0d2d37c8467ee44f0d3f393a473e55148144f18